% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.r.functions.R
\name{create_desi_obj}
\alias{create_desi_obj}
\title{Get desi objects of different resolutions or normalization methods.}
\usage{
create_desi_obj(signal_total, mass_selected, coor_selected, npcs, nor)
}
\arguments{
\item{signal_total}{dataFrame from "_signal.unlock_mass.txt" file}

\item{mass_selected}{m/z you want}

\item{coor_selected}{dataFrame from "_signal.selected.coor.txt" file}

\item{npcs}{number of principal components in RunPCA and FindNeighbors}

\item{nor}{Normalization methods include SCTransform and ScaleData.}
}
\value{
desi objects of class Seurat
}
\description{
Get desi objects of different resolutions or normalization methods.
}
\examples{
" "

}
