% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.r.functions.R
\name{create_desi_obj_with_res}
\alias{create_desi_obj_with_res}
\title{Get desi objects with the optimal number of principal components, normalization methods and resolutions.}
\usage{
create_desi_obj_with_res(
  signal_total,
  mass_selected,
  coor_selected,
  npcs,
  nor,
  res
)
}
\arguments{
\item{signal_total}{dataFrame from "_signal.unlock_mass.txt" file}

\item{mass_selected}{m/z you want}

\item{coor_selected}{dataFrame from "_signal.selected.coor.txt" file}

\item{npcs, nor, res}{the optimal number of principal components, normalization
methods and resolutions you chose from plot_seurat_clusters_with_resolution_SCT'
or 'plot_seurat_clusters_with_resolution_RNA'.}
}
\value{
desi objects of class Seurat
}
\description{
Get desi objects with the optimal number of principal components, normalization methods and resolutions.
}
\examples{
" "

}
