% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.r.functions.R
\name{Create_New_Obj}
\alias{Create_New_Obj}
\title{Get desi object from RDS file}
\usage{
Create_New_Obj(sample_name, dir_path, obj_path, dir_mass_merged)
}
\arguments{
\item{sample_name}{the file name of the sample you want to obtain}

\item{dir_path}{home directory where your '.rds' is stored}

\item{obj_path}{suffix for 'desi.lock.mass.adjust.rds' file path, subdirectory of 'dir_path'}

\item{dir_mass_merged}{directory name you want to store the results}
}
\value{
desi object of different sample that you are interested in
}
\description{
Get desi object from RDS file
}
\examples{
" "

}
