% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.r.functions.R
\name{get_lipid}
\alias{get_lipid}
\title{Get metabolites whose annotated type belongs to lipid.}
\usage{
get_lipid(colon_sub)
}
\arguments{
\item{colon_sub}{A subset of 'colon_cancer_desi.clustered_mass.table.with.anno.csv'}
}
\value{
Four matrices with different levels of screening.
        lipid_T: Lipid annotations are not empty.
        lipid_T1: Lipid annotations are not empty and are unique.
        lipid_T2: The lipid structure formula is unique.
        lipid_T3: Get the structure formula.
}
\description{
Get metabolites whose annotated type belongs to lipid.
}
\examples{
" "

}
