% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.r.functions.R
\name{plot_MASS_from_desi_obj_RNA}
\alias{plot_MASS_from_desi_obj_RNA}
\title{Plot the distribution of m/z that you are interested in.}
\usage{
plot_MASS_from_desi_obj_RNA(
  desi_obj,
  signal_total,
  mass_value,
  mass_tag,
  dir_name
)
}
\arguments{
\item{desi_obj}{the output of 'create_desi_obj_with_res'}

\item{signal_total}{dataFrame from "_signal.unlock_mass.txt" file}

\item{mass_value}{m/z value that you are interested in}

\item{mass_tag}{Set a label that attached to the file name}

\item{dir_name}{Set a dir path you want to store the results.}
}
\value{
distribution plot of m/z that you are interested in
}
\description{
Plot the distribution of m/z that you are interested in.
}
\examples{
" "

}
