% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.r.functions.R
\name{plot_seurat_clusters}
\alias{plot_seurat_clusters}
\title{Plot reconstructed tissue patterns}
\usage{
plot_seurat_clusters(desi_obj, signal_total, npcs, dir_name, col)
}
\arguments{
\item{desi_obj}{the output of 'create_desi_obj_with_res'}

\item{signal_total}{dataFrame from "_signal.unlock_mass.txt" file}

\item{npcs}{number of principal components in RunPCA and FindNeighbors}

\item{dir_name}{Set a dir path you want to store the results.}

\item{col}{type of colors}
}
\value{
reconstructed tissue patterns of all clusters and each cluster
}
\description{
Plot reconstructed tissue patterns
}
\examples{
" "

}
