% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.r.functions.R
\name{plot_seurat_clusters_with_resolution_SCT}
\alias{plot_seurat_clusters_with_resolution_SCT}
\title{Plot seurat clusters with different resolutions. The normalization method is SCTransform.}
\usage{
plot_seurat_clusters_with_resolution_SCT(
  desi_obj,
  signal_total,
  npcs,
  dir_name,
  col
)
}
\arguments{
\item{desi_obj}{the output of 'create_desi_obj'}

\item{signal_total}{dataFrame from "_signal.unlock_mass.txt" file}

\item{npcs}{number of principal components in RunPCA and FindNeighbors}

\item{dir_name}{Set a dir path you want to store the results.}

\item{col}{type of colors}
}
\value{
Cluster results and dimplots. For each fixed resolution and number
of principal components, you get a '.cluster.pdf' and a 'Dimplot.*.pdf'.
}
\description{
Plot seurat clusters with different resolutions. The normalization method is SCTransform.
}
\examples{
" "

}
